Chatbot/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
sentence-transformers
chromadb
openai
neo4j
ollama